Add[#53]: more questions and customization options to trivia game
It works by creating an array of questions, each of which has a question and an answer. The game then asks the user each question in the array and checks their answer. If the user answers correctly, their score is incremented. At the end of the game, the user's score is displayed.